Extend ServerThread optimization to test_model_registry.py and create shared utility
- Extract ServerThread class to tests/tracking/integration_test_utils.py for reusability
- Update test_model_registry.py to use ServerThread instead of subprocess-based _init_server
- Apply same ~20x performance improvement to model registry tests
- Maintain test isolation through proper backend store reset
- Update test_rest_tracking.py to import ServerThread from shared utility

Performance benefits:
- Consistent fast server startup across all tracking tests
- Reduced test execution time for model registry integration tests
- Eliminated subprocess overhead in favor of in-process threading
